+++ packages/components-react/projects/components-wrapper/src/mocks/p-textarea-wrapper-mock.jsx
@@ -2,8 +2,8 @@ const React = require('react');
 
 module.exports = {
   PTextareaWrapper: (props) => (
-    <p-textarea-wrapper label={props.label} message={props.message}>
-      {props.hideLabel !== true && props.label}
+    <p-textarea-wrapper {...props}>
+      {!props.hideLabel && props.label}
       {props.children}
       {props.state && props.state !== 'none' && props.message}
     </p-textarea-wrapper>