+++ packages/components-js/src/components/form/text-field-wrapper/text-field-wrapper.tsx
@@ -6,9 +6,10 @@ import {
   prefix,
   transitionListener,
   insertSlottedStyles,
-  randomString
+  randomString,
+  handleButtonEvent
 } from '../../../utils';
-import { FormState } from '../../../types';
+import { ButtonType, FormState } from '../../../types';
 
 @Component({
   tag: 'p-text-field-wrapper',
@@ -39,7 +40,9 @@ export class TextFieldWrapper {
   @State() private showPassword = false;
 
   private input: HTMLInputElement;
+  private searchButtonType: ButtonType = 'submit';
   private isPasswordToggleable: boolean;
+  private isInputTypeSearch: boolean;
   private labelId = randomString();
 
   public componentWillLoad(): void {
@@ -47,6 +50,7 @@ export class TextFieldWrapper {
     this.setAriaAttributes();
     this.setState();
     this.updatePasswordToggleable();
+    this.initInputTypeSearch();
     this.bindStateListener();
     this.addSlottedStyles();
   }
@@ -102,6 +106,16 @@ export class TextFieldWrapper {
           {this.isPasswordToggleable &&
           <button type='button' class={buttonClasses} onClick={(): void => this.togglePassword()} disabled={this.disabled}>
             <p-icon name={this.showPassword ? 'view-off' : 'view'} color='inherit'/>
+          </button>
+          }
+          {this.isInputTypeSearch &&
+          <button
+            onClick={(event: MouseEvent): void => this.onSubmitHandler(event)}
+            type='submit'
+            class={buttonClasses}
+            disabled={this.disabled || this.readonly}
+          >
+            <p-icon name='search' color='inherit'/>
           </button>
           }
         </div>
@@ -193,6 +207,19 @@ export class TextFieldWrapper {
     this.labelClick();
   }
 
+  private initInputTypeSearch(): void {
+    this.isInputTypeSearch = this.input.type === 'search';
+    if (this.isInputTypeSearch) {
+      this.input.style.cssText = 'padding-right: 3rem !important';
+    }
+  }
+
+  private onSubmitHandler(event: MouseEvent): void {
+    if (this.isInputTypeSearch) {
+      handleButtonEvent(event, this.host, () => this.searchButtonType, () => this.disabled);
+    }
+  }
+
   private addSlottedStyles(): void {
     const tagName = this.host.tagName.toLowerCase();
     const style = `${tagName} a {
@@ -224,7 +251,11 @@ export class TextFieldWrapper {
 
     ${tagName} input[type=password]::-webkit-contacts-auto-fill-button,
     ${tagName} input[type=password]::-webkit-credentials-auto-fill-button {
-      margin-right: 32px;
+      margin-right: 2rem;
+    }
+
+    ${tagName} input[type=search]::-webkit-search-cancel-button {
+      margin-right: 2rem;
     }
     `;