+++ packages/components-react/projects/components-wrapper/src/mocks/p-button-pure-mock.jsx
@@ -1,5 +1,7 @@
 const React = require('react');
 
 module.exports = {
-  PButtonPure: (props) => <p-button-pure onClick={props.onClick}>{props.children}</p-button-pure>
+  PButtonPure: (props) => <p-button-pure {...props}>
+    <button>{props.children}</button>
+  </p-button-pure>
 };