+++ packages/components-react/projects/components-wrapper/src/mocks/p-button-pure-mock.jsx
@@ -1,0 +1,7 @@
+const React = require('react');
+
+module.exports = {
+  PButtonPure: (props) => <p-button-pure {...props}>
+    <button>{props.children}</button>
+  </p-button-pure>
+};