+++ packages/components-react/projects/components-wrapper/src/mocks/p-flex-mock.jsx
@@ -1,0 +1,5 @@
+const React = require('react');
+
+module.exports = {
+  PFlex: (props) => <p-flex {...props}>{props.children}</p-flex>
+};