+++ packages/components-react/projects/components-wrapper/src/mocks/p-headline-mock.jsx
@@ -1,0 +1,5 @@
+const React = require('react');
+
+module.exports = {
+  PHeadline: (props) => <p-headline {...props}>{props.children}</p-headline>
+};