+++ packages/components-react/projects/components-wrapper/src/mocks/p-radio-button-wrapper-mock.jsx
@@ -1,0 +1,11 @@
+const React = require('react');
+
+module.exports = {
+  PRadioButtonWrapper: (props) => (
+    <p-radio-button-wrapper {...props}>
+      {!props.hideLabel && props.label}
+      {props.children}
+      {props.state && props.state !== 'none' && props.message}
+    </p-radio-button-wrapper>
+  ),
+};