+++ packages/components-js/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Search submit button to `p-text-field-wrapper` type search component
+
 ### Changed
 - Visual appearance of icons
 - Improve caching strategy for icons by content-based hash