+++ packages/components-js/CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Submit button with search icon to `p-textfield-wrapper` type search
+
 ### Changed
+- Visual appearance of icons
+- Improve caching strategy for icons by content-based hash
 - Cursor of Radio, Checkbox and Select
 
 ## [1.4.0-rc.0] - 2020-04-09