+++ packages/components-js/CHANGELOG.md
@@ -8,10 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Description property to `p-text-field-wrapper`, `p-textarea-wrapper` and `p-select-wrapper` 
+- Submit button with search icon to `p-textfield-wrapper` type search
 
 ### Changed
 - Improve accessibility of error and success states of form elements
 - Aria-invalid attribute of form elements if they are in error state is now managed by component
+- Visual appearance of icons
+- Improve caching strategy for icons by content-based hash
+- Cursor of Radio, Checkbox and Select
 
 ## [1.4.0-rc.0] - 2020-04-09